rename file admin_filter to filters
because it contains many filters, not only admin
