+++ core/filters/filters.py
@@ -70,7 +70,7 @@ class HashTagFilter(BaseFilter):
 
 class StrictChatFilter(BaseFilter):
     """
-    Отделяет чаты, в которых должен работать хэндлер от тех, где не должен
+    Отделяет чаты, в которых должен работать стрикт мод от тех, где не должен
     """
     async def __call__(self, message: Message, chat_settings: dict) -> bool:
         strict_chats = chat_settings.get('strict_chats', [])
